reorganize project framework
