--- PI/pid.c
@@ -1,56 +1,0 @@
-#include "pid.h"
-
-/**
- * @brief Initialize pid controller
- *  kp and ki are the parameter generated by s transfer function
- *  Transfer function : 
- *                    ki
- *      G(s) = kp + ------
- *                    s
- *  Actually, the corresponding transfer function in z domain:
- *                      ki * z
- *      G(z) = kp + ---------------
- *                      z - 1 
- * 
- * @param s Pointer to pidStruct
- * @param kp Value of kp
- * @param ki Value of ki 
- * @param ts Value of sampling period
- */
-void pid_init(pidStruct * s, float kp, float ki, float ts) {
-    s->kp = kp ;
-    s->ts = ts ;
-    s->ki = ki * ts ;
-    s->sum_ki = 0.0f ;
-    s->output = 0.0f ;
-    s->reference = 0.0f ;
-}
-
-
-/**
- * @brief Update reference dynamically if needed
- * 
- * @param s Pointer to pidStruct
- * @param ref Value of reference to be set 
- */
-void pid_update_ref(pidStruct * s, float ref){
-    s->reference = ref ;
-}
-
-
-/**
- * @brief 
- * PI controller is really simple. 
- * You can implement anti-saturation or other features by yourself.
- * pid_calc should be called every control period.
- * 
- * @param s Pointer to pidStruct
- * @param feedback Value of feedback signal
- * @return float Output of PI controller
- */
-float pid_calc(pidStruct * s, float feedback) {
-    float error = s->reference - feedback ;
-    s->sum_ki += s->ki * error ;
-    s->output = s->sum_ki + s->kp * error ;
-    return s->output ;
-}
