--- PR/test/README.md
@@ -1,14 +1,0 @@
-# Test
-
-1. compile test program and run:
-
-    ```bash
-    make test
-    ./test_pr.elf
-    ```
-
-2. run test_pr_plot.m by Octave/MATLAB to see results:
-
-![Test results](test_result.jpg)
-
-As can be seen, a sine wave with an amplitude of 1 is amplified to a amplitude of 10, corresponding to the resonant gain of 10.
