--- RC/pwrc/test/test_result_plot.m
@@ -1,5 +1,0 @@
-clear; 
-close all ;
-
-d = load('text_output.txt') ;
-plot(d(:,1));
