+++ examples/RC/FORC/test_result_plot.m
@@ -1,0 +1,5 @@
+clear; 
+close all ;
+
+d = load('text_output.txt') ;
+plot(d(:,1));
