--- filter/test/README.md
@@ -1,14 +1,0 @@
-# Test
-
-1. compile test program and run:
-
-    ```bash
-    make test
-    ./test_filter.elf
-    ```
-
-2. run test_filter_plot.m by Octave/MATLAB to see results:
-
-![Test results](test_results.jpg)
-
-As can be seen, results computed by C program and Octave are the same.
